Modularize trust system
So the files can share the same logic
